DRY: Introduce File_Loader_Provider_Trait and improve doc for tests
